Towards renaming of powertorch fluxes from, e.g., Pe to QeMWm2  for much more clarity
